fashion-stylist-app/requirements.txt
tensorflow
gradio
Pillow
